started on batch 4, added milestone for prog 1

+++ Batch_3_Codes/Exempted_Duplicates.py
@@ -5,7 +5,7 @@ numbers = []
 print("Please input 10 numbers")
 
 for i in range(10):
-    num = int(input(f"Enter number {i+1}: "))
+    num = float(input(f"Enter number {i+1}: "))
     numbers.append(num)
 
 #List for numbers without duplicate and first instance of duplicate numbers
